\name{CompDb.Hsapiens.HMDB.5.0}
\docType{package}

\alias{CompDb.Hsapiens.HMDB.5.0-package}
\alias{CompDb.Hsapiens.HMDB.5.0}
\alias{CompDb.Hsapiens.HMDB.5.0}


\title{HMDB compound annotation package}

\description{
  This package contains a \code{\link[CompoundDb]{CompDb}} database
  providing compound annotations based on HMDB (version 5.0).
  For examples and help on functions see the help pages from the
  \code{CompoundDb} package.
}

\author{A Gautam}

\examples{
## load the library
library(CompDb.Hsapiens.HMDB.5.0)
## list the contents that are loaded into memory
ls('package:CompDb.Hsapiens.HMDB.5.0')
## show the db object that is loaded by calling it's name
CompDb.Hsapiens.HMDB.5.0

## for more examples see the CompoundDb package.


}

\keyword{package}
\keyword{data}
